update StreamOverview
- add TitleCell
- hide description
- update styles
