+++ graylog2-web-interface/src/components/streams/StreamsOverview/Constants.ts
@@ -21,8 +21,8 @@ export const ENTITY_TABLE_ID = 'streams';
 export const DEFAULT_LAYOUT = {
   pageSize: 20,
   sort: { attributeId: 'title', direction: 'asc' } as Sort,
-  displayedColumns: ['title', 'description', 'index_set_title', 'rules', 'throughput', 'disabled'],
-  columnsOrder: ['title', 'description', 'index_set_title', 'rules', 'throughput', 'disabled', 'created_at'],
+  displayedColumns: ['title', 'index_set_title', 'rules', 'throughput', 'disabled'],
+  columnsOrder: ['title', 'index_set_title', 'rules', 'throughput', 'disabled', 'created_at'],
 };
 
 export const ADDITIONAL_ATTRIBUTES = [
